Reorganize the ptb stuff
