+++ crates/sui/src/lib.rs
@@ -9,7 +9,7 @@ pub mod genesis_ceremony;
 pub mod genesis_inspector;
 pub mod key_identity;
 pub mod keytool;
-pub mod ptb_parser;
+pub mod ptb;
 pub mod shell;
 pub mod sui_commands;
 pub mod validator_commands;
